[Core/Player] Change money from uint32 to uint64

+++ src/server/game/Handlers/ItemHandler.cpp
@@ -517,7 +517,7 @@ void WorldSession::HandleSellItemOpcode(WorldPacket& recvData)
         {
             if (pProto->SellPrice > 0)
             {
-                uint32 money = pProto->SellPrice * count;
+                uint64 money = uint64(pProto->SellPrice) * count;
                 if (count < pItem->GetCount())               // need split items
                 {
                     Item* pNewItem = pItem->CloneItem(count, _player);
